requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
bson==0.5.10
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.0
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.39.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
multidict==6.6.4
narwhals==2.5.0
numpy==2.0.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-geohash==0.8.5
pytz==2025.2
requests==2.32.5
rsa==4.9.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
walrus-python==0.1.0
-e git+https://github.com/surfer05/btchd@52f2ffcefe6f48f80803fb76e2ea0ff914e7a36c#egg=walrusdb&subdirectory=walrusdb
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
